refactor(MenuValidator): Price 이관으로 인한 수정
